+++ Frends.Excel.ConvertToJSON/Frends.Excel.ConvertToJSON.Tests/ErrorHandlerTest.cs
@@ -1,0 +1,51 @@
+using System;
+using System.IO;
+using Frends.Excel.ConvertToJSON.Definitions;
+using NUnit.Framework;
+
+namespace Frends.Excel.ConvertToJSON.Tests;
+
+[TestFixture]
+public class ErrorHandlerTest
+{
+    private const string CustomErrorMessage = "CustomErrorMessage";
+
+    private static Input DefaultInvalidInput() =>
+        new() { Path = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, Guid.NewGuid() + ".xlsx") };
+
+    [Test]
+    public void Should_Throw_Error_When_ThrowErrorOnFailure_Is_True()
+    {
+        var options = new Options { ThrowErrorOnFailure = true };
+
+        Assert.That(
+            () => Excel.ConvertToJSON(DefaultInvalidInput(), options, default),
+            Throws.TypeOf<FileNotFoundException>());
+    }
+
+    [Test]
+    public void Should_Return_Failed_Result_When_ThrowErrorOnFailure_Is_False()
+    {
+        var options = new Options { ThrowErrorOnFailure = false };
+
+        var result = Excel.ConvertToJSON(DefaultInvalidInput(), options, default);
+
+        Assert.That(result.Success, Is.False);
+        Assert.That(result.Error, Is.Not.Null);
+    }
+
+    [Test]
+    public void Should_Use_Custom_ErrorMessageOnFailure()
+    {
+        var options = new Options
+        {
+            ThrowErrorOnFailure = true,
+            ErrorMessageOnFailure = CustomErrorMessage,
+        };
+
+        var ex = Assert.Throws<Exception>(() => Excel.ConvertToJSON(DefaultInvalidInput(), options, default));
+
+        Assert.That(ex, Is.Not.Null);
+        Assert.That(ex!.Message, Contains.Substring(CustomErrorMessage));
+    }
+}