+++ Frends.Excel.ConvertToJSON/Frends.Excel.ConvertToJSON/Definitions/Options.cs
@@ -13,21 +13,29 @@ public class Options
     /// </summary>
     /// <example>Sheet1</example>
     [DefaultValue(@"")]
-    public string ReadOnlyWorkSheetWithName { get; set; } = "";
+    public string ReadOnlyWorkSheetWithName { get; set; } = string.Empty;
 
     /// <summary>
     /// If set to true, numbers will be used as column headers instead of letters (A = 1, B = 2...).
     /// </summary>
     /// <example>true</example>
-    [DefaultValue("false")]
-    public bool UseNumbersAsColumnHeaders { get; set; }
+    [DefaultValue(false)]
+    public bool UseNumbersAsColumnHeaders { get; set; } = false;
 
     /// <summary>
     /// Choose if exception should be thrown when conversion fails.
     /// </summary>
     /// <example>true</example>
-    [DefaultValue("true")]
-    public bool ThrowErrorOnFailure { get; set; }
+    [DefaultValue(true)]
+    public bool ThrowErrorOnFailure { get; set; } = true;
+
+    /// <summary>
+    /// Overrides the error message on failure.
+    /// </summary>
+    /// <example>Excel conversion failed: input file was invalid</example>
+    [DisplayFormat(DataFormatString = "Text")]
+    [DefaultValue("")]
+    public string ErrorMessageOnFailure { get; set; } = string.Empty;
 
     /// <summary>
     /// Date format selection.
@@ -43,6 +51,6 @@ public class Options
     /// Default false
     /// </summary>
     /// <example>true</example>
-    [DefaultValue("false")]
-    public bool ShortDatePattern { get; set; }
+    [DefaultValue(false)]
+    public bool ShortDatePattern { get; set; } = false;
 }